Moved binding logic into commands module
