+++ src/commands/bindings.rs
